# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Akamai CCM API
  version: 1.0.0
  description: >
    Cloud Compliance Manager API for managing Linode accounts, resources,
    compliance rules, reports, and users.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://<your-host>
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, power_user, auditor]
        can_view_costs:
          type: boolean
        can_view_compliance:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserAccess:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        account_name:
          type: string
        granted_by:
          type: string
          format: uuid
        can_view_costs:
          type: boolean
        can_view_compliance:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List all users
      description: Requires `admin` role.
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create a user
      description: Requires `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                full_name:
                  type: string
                  default: ''
                role:
                  type: string
                  enum: [admin, power_user, auditor]
                  default: auditor
                can_view_costs:
                  type: boolean
                  default: true
                can_view_compliance:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Users]
      summary: Update a user
      description: Requires `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                role:
                  type: string
                  enum: [admin, power_user, auditor]
                is_active:
                  type: boolean
                can_view_costs:
                  type: boolean
                can_view_compliance:
                  type: boolean
                password:
                  type: string
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [Users]
      summary: Delete a user
      description: Requires `admin` role.
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
  /api/users/{user_id}/access:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Get user account access
      description: Requires `admin` role.
      responses:
        '200':
          description: List of access grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserAccess'
    post:
      tags: [Users]
      summary: Grant account access to user
      description: Requires `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id]
              properties:
                account_id:
                  type: string
                  format: uuid
                can_view_costs:
                  type: boolean
                  default: true
                can_view_compliance:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Access granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  granted:
                    type: boolean
  /api/users/{user_id}/access/{account_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Users]
      summary: Revoke user account access
      description: Requires `admin` role.
      responses:
        '200':
          description: Access revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: boolean

  # ─── Refresh / Sync ───────────────────────────────────────────────────────
...